requirements.txt
sentence-transformers
faiss-cpu
transformers
torch
fastmcp
PyMuPDF
numpy
accelerate